backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.28.0
python-dotenv==1.2.1
httpx>=0.27.0
langchain==1.2.10
langchain-core==1.2.15
langchain-anthropic>=1.4.0
langgraph==1.0.9
pytest>=9.0.0
pytest-asyncio>=1.3.0
